+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.12.0 (Unreleased)
+## 0.12.0 (2026-09-09)
 
 ### New Features